Rename package to ProcessRunner
