+++ Sources/ProcessRunner/ProcessResult.swift
